requirements_rasa.txt
rasa==3.6.21
rasa-sdk==3.6.2
requests==2.32.5
psycopg2-binary==2.9.11
SQLAlchemy==1.4.54
